requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
selenium==4.18.1
webdriver-manager==4.0.1
aiofiles==23.2.1
python-multipart==0.0.9
openai==1.12.0
unrealspeech==0.1.0
pygame==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-magic==0.4.27
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1 